DeviantCord 4 Public Release V4.0.9
The previous versions, were significantly broken, during the testing period to make sure that we were at a point that others could start contributing.

I reworked a lot of the main file, but there is still work to be done.

I will upload those versions that are left. Some of them are just not around due to rapid iteration.